refactor: removeSectionByStation의 책임을 sectinos로 이동
